Migrate frontend from Stacks to Stellar/Soroban architecture
Closes #201: Rename StacksProvider and related hooks to chain-agnostic names
- StacksProvider -> WalletProvider
- useStacks -> useWallet
- Backward compatibility maintained via re-exports

Closes #204: Migrate predinex-read-api away from Hiro/Stacks Core
- Remove Hiro-specific API functions
- Consolidate to Soroban-native reads

Closes #205: Deprecate stacks-api.ts
- Add deprecation notice with migration path
- Point to soroban-read-api.ts as canonical source

Closes #211: Replace STX-specific amount helpers
- Remove STX/XLM-specific aliases
- Keep asset-agnostic formatting functions
- Use configurable TOKEN_SYMBOL from config
